pyproject.toml
[tool.poetry]
name = "jobfunnel"
version = "3.0.2"
description = "Scrape job websites into a single spreadsheet with no duplicates."
authors = ["anoduck <11767-anoduck@users.noreply.gitgud.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = ">=4.6.3"
lxml = ">=4.2.4"
requests = ">=2.19.1"
python-dateutil = ">=2.8.0"
pyyaml = ">=5.1"
nltk = ">=3.4.1"
pytest-mock = ">=3.1.1"
selenium = ">=3.141.0"
webdriver-manager = ">=2.4.0"
cerberus = ">=1.3.2"
tqdm = ">=4.47.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
